MADE SHIFTABLE MODULE! with shift method that takes two arguments: key and today
