+++ lib/shift.rb
@@ -1,45 +1,20 @@
+require './lib/shiftable'
+
 class Shift
+  include Shiftable
 
-  def date_squared(today)
-    today.to_i ** 2
-  end
-
-  def date_squared_last(date_squared)
-    date_squared.digits[0..3].reverse
-  end
-
-  def offsets(today)
-    date_squared_last(date_squared(today))
-  end
-
-  def all_keys(key)
-    keys = []
-    key.split("").each_cons(2) do |num|
-      keys << num[0] + num[1]
-    end
-    keys.map { |k| k.to_i }
-  end
-
-  def shifts(all_keys)
-    zips = []
-    all_keys.zip(offsets(today)) do |key|
-      zips << key.sum
-    end
-    zips
-  end
-
-  def shifts_by_name
+def shifts_by_name
     names = ["A", "B", "C", "D"]
-    two_d = names.zip(shifts(all_keys(key)))
+    two_d = names.zip(shifts(key, today))
       two_d.reduce({}) do |hash, key|
         hash[key[0]] = key[1]
         hash
     end
   end
 
-  def ordinal_values
+  def ordinal_values(message)
     ords = []
-      @message.chars.map do |char|
+      message.chars.map do |char|
         if char.ord != 32
         ords << char.downcase.ord
       else
